Generate en_US.UTF-8 locale
Even though the LANG environment variable is set to LANG=en_US.UTF-8 by
heroku buildpacks, the locale has not been generated (see the output of
locale -a). In the case of a Ruby app, all strings are defaulted to
US-ASCII.

This commit will generate the en_US.UTF-8 locale.